requirements.txt
asgiref==3.4.1
attrs==21.2.0
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-spectacular==0.17.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==3.2.0
model-bakery==1.3.2
psycopg2==2.9.1
PyJWT==2.1.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
whitenoise==5.2.0
channels~=3.0.4